telegram_adapter/requirements.txt
python-telegram-bot>=21.0,<22.0
httpx>=0.27

